AXS-2833: Relax Start Solution Unittest
This did rely on the fact, that an optimal start solution would always
be returned as the optimal solution after the solve.
This did not work with HiGHS and there is no general reason it should.
So instead, we rely on the fact that the implementations
throw on infeasible start solutions.